# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "APPROVE",
              "RETURN"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "action"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InspectionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "allocation_id": {
            "format": "uuid",
            "title": "Allocation Id",
            "type": "string"
          },
          "extraction_run_id": {
            "format": "uuid",
            "title": "Extraction Run Id",
            "type": "string"
          }
        },
        "required": [
          "allocation_id",
          "extraction_run_id"
        ],
        "title": "InspectionCreateRequest",
        "type": "object"
      },
      "InspectionResponse": {
        "additionalProperties": false,
        "properties": {
          "allocation_id": {
            "format": "uuid",
            "title": "Allocation Id",
            "type": "string"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "candidate_decision": {
            "title": "Candidate Decision",
            "type": "string"
          },
          "final_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Decision"
          },
          "inspection_id": {
            "format": "uuid",
            "title": "Inspection Id",
            "type": "string"
          },
          "judgments": {
            "items": {
              "$ref": "#/components/schemas/JudgmentView"
            },
            "title": "Judgments",
            "type": "array"
          },
          "material_lot_id": {
            "format": "uuid",
            "title": "Material Lot Id",
            "type": "string"
          },
          "revision_no": {
            "title": "Revision No",
            "type": "integer"
          },
          "round_no": {
            "title": "Round No",
            "type": "integer"
          },
          "spec_snapshot": {
            "additionalProperties": true,
            "title": "Spec Snapshot",
            "type": "object"
          },
          "spec_version_id": {
            "format": "uuid",
            "title": "Spec Version Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "inspection_id",
          "material_lot_id",
          "allocation_id",
          "spec_version_id",
          "spec_snapshot",
          "status",
          "candidate_decision",
          "final_decision",
          "version",
          "round_no",
          "revision_no",
          "blockers",
          "judgments"
        ],
        "title": "InspectionResponse",
        "type": "object"
      },
      "InspectionReturnRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "target_spec_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Spec Item Id"
          }
        },
        "required": [
          "reason"
        ],
        "title": "InspectionReturnRequest",
        "type": "object"
      },
...
      "InternalResultItem": {
        "additionalProperties": false,
        "properties": {
          "spec_item_id": {
            "format": "uuid",
            "title": "Spec Item Id",
            "type": "string"
          },
          "values": {
            "items": {
              "pattern": "^-?(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "spec_item_id",
          "values"
        ],
        "title": "InternalResultItem",
        "type": "object"
      },
      "InternalResultsRequest": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/InternalResultItem"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "InternalResultsRequest",
        "type": "object"
      },
      "JudgmentView": {
        "additionalProperties": false,
        "properties": {
          "effective_decision": {
            "title": "Effective Decision",
            "type": "string"
          },
          "hyc_reference_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hyc Reference Decision"
          },
          "internal_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Decision"
          },
          "item_code": {
            "title": "Item Code",
            "type": "string"
          },
          "spec_item_id": {
            "format": "uuid",
            "title": "Spec Item Id",
            "type": "string"
          },
          "supplier_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Decision"
          }
        },
        "required": [
          "spec_item_id",
          "item_code",
          "supplier_decision",
          "hyc_reference_decision",
          "internal_decision",
          "effective_decision"
        ],
        "title": "JudgmentView",
        "type": "object"
      },
      "LineageRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "LineageRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "HYC Inspection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inspections": {
      "post": {
        "operationId": "create_case_api_v1_inspections_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Case Api V1 Inspections Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}": {
      "get": {
        "operationId": "get_case_api_v1_inspections__inspection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/approvals": {
      "post": {
        "operationId": "approval_api_v1_inspections__inspection_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-P3-Fault-At",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-P3-Fault-At"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approval",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/internal-results": {
      "put": {
        "operationId": "internal_results_api_v1_inspections__inspection_id__internal_results_put",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalResultsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Internal Results",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/retests": {
      "post": {
        "operationId": "retest_api_v1_inspections__inspection_id__retests_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retest",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/return": {
      "post": {
        "operationId": "return_case_api_v1_inspections__inspection_id__return_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Return Case",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/revisions": {
      "post": {
        "operationId": "revision_api_v1_inspections__inspection_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision",
        "tags": [
          "p3-inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/submit": {
      "post": {
        "operationId": "submit_case_api_v1_inspections__inspection_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Case",
        "tags": [
          "p3-inspections"
        ]
      }
    },
...
  }
}